refactor: compute CourseSchedule status client-side from datetime
Extract table rendering into a new CourseScheduleTable client component
so status (開放/關閉/結束) is computed from Date.now() vs start/end_time
at render time, bypassing the stale DB status field under "use cache".
Also serialize Mongoose ObjectId to string to fix the toJSON prop error,
and use a manual UTC+8 formatter to avoid hydration mismatch from
locale-dependent toLocaleString output.
